Fixed Contract Bugs; Implemented NFT Minting Functionality; Implemented Video Upload Functionality

+++ frontend/pages/creators/index.js
@@ -90,7 +90,7 @@ const Creators = () => {
                   ethers.utils
                     .formatEther(
                       (
-                        parseFloat(
+                        parseInt(
                           ethers.BigNumber.from(
                             ethers.utils.parseEther(data[2].toString())
                           ).toString()
